chore(api): dedup enemyIdFromSlug + drop dead config.site field
Post-review cleanup of the v0.59.0 public API surface (ponytail-review,
adversarially verified, high-confidence only):

- enemyIdFromSlug was copy-pasted identically into three projection
  modules (status/map/stats), each just `FACTION_INDEX[enemy]`. Define it
  once in enums/factions.mjs (next to FACTION_INDEX/FACTION_SLUG_BY_ID);
  routes import from there. Three duplicate test blocks collapse into one
  in factions.test.mjs.
- Remove the never-read `site` field from the parsed server config; every
  consumer imports SITE_URL from config/site.mjs directly.

No behavior change. Churn on code new to this unreleased milestone, so no
version bump or changelog entry.
