+++ src/app/api/v1/h1/stats/statsProjection.mjs
@@ -1,5 +1,5 @@
 import { z } from 'zod';
-import { FACTION_SLUG_BY_ID, FACTION_INDEX } from '@/shared/enums/factions.mjs';
+import { FACTION_SLUG_BY_ID } from '@/shared/enums/factions.mjs';
 import { encodeCursor } from '@/shared/utils/api/cursor.mjs';
 
 /**
@@ -36,14 +36,6 @@ export function parseStatsQuery(searchParams) {
         };
     }
     return { success: true, data: result.data };
-}
-
-/**
- * @param {string} enemy - Faction slug.
- * @returns {number} faction id.
- */
-export function enemyIdFromSlug(enemy) {
-    return FACTION_INDEX[enemy];
 }
 
 /**
